Fix SimpleOverlayArea
