+++ modules/lab/src/main/java/org/locationtech/jts/operation/overlayarea/OverlayArea.java
@@ -11,11 +11,8 @@
  */
 package org.locationtech.jts.operation.overlayarea;
 
-import org.locationtech.jts.algorithm.Angle;
 import org.locationtech.jts.algorithm.Area;
-import org.locationtech.jts.algorithm.LineIntersector;
 import org.locationtech.jts.algorithm.Orientation;
-import org.locationtech.jts.algorithm.RobustLineIntersector;
 import org.locationtech.jts.algorithm.locate.IndexedPointInAreaLocator;
 import org.locationtech.jts.algorithm.locate.PointOnGeometryLocator;
 import org.locationtech.jts.algorithm.locate.SimplePointInAreaLocator;
@@ -32,7 +29,6 @@ import org.locationtech.jts.index.kdtree.KdTree;
 import org.locationtech.jts.math.MathUtil;
 import org.locationtech.jts.noding.BasicSegmentString;
 import org.locationtech.jts.noding.MCIndexSegmentSetMutualIntersector;
-import org.locationtech.jts.noding.SegmentIntersector;
 import org.locationtech.jts.noding.SegmentSetMutualIntersector;
 import org.locationtech.jts.noding.SegmentString;
 
@@ -43,7 +39,7 @@ import java.util.List;
  * Computes the area of the overlay of two polygons without forming
  * the actual topology of the overlay.
  * Since the topology is not needed, the computation is
- * is insensitive to the fine details of the overlay topology,
+ * insensitive to the fine details of the overlay topology,
  * and hence is fully robust.
  * It also allows for a simpler implementation with more aggressive
  * performance optimization.
@@ -78,8 +74,6 @@ public class OverlayArea {
     return geom0.getEnvelopeInternal().intersects(geom1.getEnvelopeInternal());
   }
 
-  private static LineIntersector li = new RobustLineIntersector();
-  
   private Geometry geom0;
   private Envelope geomEnv0;
   private IndexedPointInAreaLocator locator0;
@@ -220,117 +214,6 @@ public class OverlayArea {
     return intVisitor.getArea();
   }
 
-  class IntersectionVisitor implements SegmentIntersector {
-    private double area = 0.0;
-
-    double getArea() {
-      return area;
-    }
-
-    @Override
-    public void processIntersections(SegmentString a, int aIndex, SegmentString b, int bIndex) {
-      boolean isCCWA = (boolean) a.getData();
-      boolean isCCWB = (boolean) b.getData();
-
-      Coordinate a0 = a.getCoordinate(aIndex);
-      Coordinate a1 = a.getCoordinate(aIndex + 1);
-      Coordinate b0 = b.getCoordinate(bIndex);
-      Coordinate b1 = b.getCoordinate(bIndex + 1);
-
-      if (isCCWA) {
-        Coordinate tmp = a0; a0 = a1; a1 = tmp;
-      }
-      if (isCCWB) {
-          Coordinate tmp = b0; b0 = b1; b1 = tmp;
-      }
-
-      li.computeIntersection(a0, a1, b0, b1);
-      if (! li.hasIntersection()) return;
-
-      Coordinate intPt = li.getIntersection(0);
-
-      if (li.isProper() || li.isInteriorIntersection()) {
-        // Edge-edge intersection OR vertex-edge intersection
-
-        /**
-         * An intersection creates two edge vectors which contribute to the area.
-         *
-         * With both rings oriented CW (effectively)
-         * There are two situations for segment intersection:
-         *
-         * 1) A entering B, B exiting A => rays are IP->A1:R, IP->B0:L
-         * 2) A exiting B, B entering A => rays are IP->A0:L, IP->B1:R
-         * (where IP is the intersection point,
-         * and  :L/R indicates result polygon interior is to the Left or Right).
-         *
-         * For accuracy the full edge is used to provide the direction vector.
-         */
-
-        if (Orientation.CLOCKWISE == Orientation.index(a0, a1, b0)) {
-          if (intPt.equals2D(a1)) {
-            // Intersection at vertex and A0 -> A1 is outside the intersection area.
-            // Area will be computed by the segment A1 -> A2
-            return;
-          }
-          area += EdgeVector.area2Term(intPt, a0, a1, true);
-          area += EdgeVector.area2Term(intPt, b1, b0, false);
-        } else if (Orientation.CLOCKWISE == Orientation.index(a0, a1, b1)) {
-          if (intPt.equals2D(a0)) {
-            // Intersection at vertex and A0 -> A1 is outside the intersection area.
-            // Area will be computed by the segment A(-1) -> A0
-            return;
-          }
-          area += EdgeVector.area2Term(intPt, a1, a0, false);
-          area += EdgeVector.area2Term(intPt, b0, b1, true);
-        }
-
-      } else {
-        // vertex-vertex intersection
-        // This intersection is visited 4 times - include only once
-        if (!a1.equals2D(b1)) {
-          return;
-        }
-
-        // If A0->A1 is collinear with B0->B1, then the intersection point might not be equal to A1 and B1
-        intPt = a1;
-
-        Coordinate a2 = a.nextInRing(aIndex + 1);
-        Coordinate b2 = b.nextInRing(bIndex + 1);
-        if (isCCWA) {
-          a2 = a.prevInRing(aIndex);
-        }
-        if (isCCWB) {
-          b2 = b.prevInRing(bIndex);
-        }
-
-        double aAngle = Angle.interiorAngle(a0, intPt, a2);
-        double bAngle = Angle.interiorAngle(b0, intPt, b2);
-
-        // The LTE ja LT are chosen such that when A0->A1 is collinear with B0->B1,
-        // or when A1->A2 is collinear with B1->B2, then A is chosen.
-        // This avoids double counting in the case of collinear segments.
-        if (Angle.interiorAngle(a0, intPt, b2) <= bAngle) {
-          area += EdgeVector.area2Term(intPt, a1, a0, false);
-        }
-        if (Angle.interiorAngle(b0, intPt, a2) <= bAngle) {
-          area += EdgeVector.area2Term(intPt, a1, a2, true);
-        }
-        if (Angle.interiorAngle(b0, intPt, a2) < aAngle) {
-          area += EdgeVector.area2Term(intPt, b1, b0, false);
-        }
-        if (Angle.interiorAngle(a0, intPt, b2) < aAngle) {
-          area += EdgeVector.area2Term(intPt, b1, b2, true);
-        }
-      }
-    }
-
-    @Override
-    public boolean isDone() {
-      // Process all intersections
-      return false;
-    }
-  }
-  
 
   private double areaForInteriorVertices(LinearRing ring) {
     /**
